+++ docs/second_page.md
@@ -1,0 +1,2 @@
+# Second page
+this is a second page
