+++ go.mod
@@ -3,7 +3,7 @@ module cutedoc
 go 1.16
 
 require (
-	github.com/stretchr/testify v1.7.0 // indirect
+	github.com/alecthomas/kong v0.2.18
 	github.com/yuin/goldmark v1.4.2
 	gopkg.in/ini.v1 v1.63.2
 )
